# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Buddies API
  version: 1.0.0
  description: |
    Dual-sided B2B reverse-auction marketplace API (consumer restaurants + suppliers).
    Auth: `Authorization: Bearer <supabase_jwt>` or dev `Bearer dev:<role>:<userId>`.
servers:
  - url: http://localhost:8000/v1
...
paths:
...
  /orders/{id}/chat/messages:
    post:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Chat message }
...
  /return-claims/{id}/chat:
    get:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Per-claim chat thread + messages }
  /return-claims/{id}/chat/messages:
    post:
      tags: [shared]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Return chat message }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...